[sailors] more cleanup and refactoring + fixes

+++ src/libs/sailors/src/sailors_menu.cpp
@@ -86,13 +86,13 @@ void Menu::Draw(VDX9RENDER *rs, SailorsPoints &sailorsPoints) const
         if (sailorsPoints.links.selected >= 0 && sailorsPoints.links.selected < sailorsPoints.links.count)
         {
             std::string s1 = "point 1= ";
-            s1 += static_cast<int32_t>(sailorsPoints.links.link[sailorsPoints.links.selected].first);
+            s1 += sailorsPoints.links.link[sailorsPoints.links.selected].first;
             s1 += "('F2')";
 
             rs->Print(rs->GetCurFont(), D3DCOLOR_XRGB(200, 200, 200), 10, 140, "%s", s1.c_str());
 
             s1 = "point 2= ";
-            s1 += static_cast<int32_t>(sailorsPoints.links.link[sailorsPoints.links.selected].next);
+            s1 += sailorsPoints.links.link[sailorsPoints.links.selected].next;
             s1 += "('F3')";
 
             rs->Print(rs->GetCurFont(), D3DCOLOR_XRGB(200, 200, 200), 10, 160, "%s", s1.c_str());
@@ -123,9 +123,9 @@ void Menu::Draw(VDX9RENDER *rs, SailorsPoints &sailorsPoints) const
         }
 
     std::string s1 = "F10 - Reload cannons(";
-    s1 += static_cast<int32_t>(reloaded);
+    s1 += reloaded;
     s1 += "/";
-    s1 += static_cast<int32_t>(count);
+    s1 += count;
     s1 += ")";
 
     rs->Print(rs->GetCurFont(), D3DCOLOR_XRGB(200, 200, 200), 10, 470, "%s", s1.c_str());
@@ -148,15 +148,15 @@ void Menu::Update(SailorsPoints &sailorsPoints)
     line[0] += ">";
 
     line[1] = "selected point <";
-    line[1] += static_cast<int32_t>(sailorsPoints.points.selected);
+    line[1] += sailorsPoints.points.selected;
     line[1] += ">(";
-    line[1] += static_cast<int32_t>(sailorsPoints.points.count);
+    line[1] += sailorsPoints.points.count;
     line[1] += ")";
 
     line[2] = "selected path <";
-    line[2] += static_cast<int32_t>(sailorsPoints.links.selected);
+    line[2] += sailorsPoints.links.selected;
     line[2] += ">(";
-    line[2] += static_cast<int32_t>(sailorsPoints.links.count);
+    line[2] += sailorsPoints.links.count;
     line[2] += ")";
 };
